Refactor LogMessageTest* and LogMessageArchiveTest* to Jupiter
