+++ src/test/java/de/konfidas/ttc/messages/LogMessageTestValidationSuccessful.java
@@ -1,5 +1,6 @@
-package de.konfidas.ttc.tars;
+package de.konfidas.ttc.messages;
 
+import de.konfidas.ttc.tars.LogMessageArchiveImplementation;
 import de.konfidas.ttc.validation.AggregatedValidator;
 import de.konfidas.ttc.validation.CertificateFileNameValidator;
 import de.konfidas.ttc.validation.LogMessageSignatureValidator;
@@ -15,39 +16,41 @@ import java.io.File;
 import java.security.Security;
 import java.util.stream.Stream;
 
-import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
 
-public class LogMessageArchiveTestValidationFailing {
+public class LogMessageTestValidationSuccessful {
     final static Logger logger = LoggerFactory.getLogger(Logger.ROOT_LOGGER_NAME);
-    final static File brokenTarFiles = new File("testdata" + File.separator + "negative" + File.separator + "validation_errors");
+    final static File correctLogs = new File("testdata" + File.separator + "positive" + File.separator + "can_validate");
 
     @BeforeEach
     public void initialize() {
         Security.addProvider(new BouncyCastleProvider());
     }
 
+
     @ParameterizedTest
     @MethodSource("filesToTest")
-    public void validateTARwithErrors_shouldShowValidationErrors(File tarFile) throws Exception {
+    public void parseTAR_shouldNotThrowException(File tarFile) throws Exception {
         logger.debug("");
         logger.debug("============================================================================");
-        logger.debug("testing tar file {}:", tarFile.getName());
+        logger.debug("parsing log message {}:", tarFile.getName());
 
-        assertFalse(((Validator) new AggregatedValidator()
+        assertTrue(((Validator) new AggregatedValidator()
                 .add(new CertificateFileNameValidator())
                 .add(new LogMessageSignatureValidator())).validate(new LogMessageArchiveImplementation(tarFile)).getValidationErrors().isEmpty());
+
     }
 
     static Stream<File> filesToTest() {
-        logger.debug("checking for Tars in " + brokenTarFiles.getName());
-        if (!brokenTarFiles.isDirectory()) {
-            fail(brokenTarFiles.getAbsolutePath() + " is not a directory.");
+        logger.debug("checking for Tars in " + correctLogs.getName());
+        if (!correctLogs.isDirectory()) {
+            fail(correctLogs.getAbsolutePath() + " is not a directory.");
         }
-        if (brokenTarFiles.listFiles() == null) {
-            fail("The directory of test TAR files is empty in: " + brokenTarFiles.getAbsolutePath());
+        if (correctLogs.listFiles() == null) {
+            fail("The directory of test TAR files is empty in: " + correctLogs.getAbsolutePath());
         }
-        return Stream.of(brokenTarFiles.listFiles());
+        return Stream.of(correctLogs.listFiles());
     }
 }
